backend/apps/films_api/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Nikita Smirnov", "Pavel Eroshkin", "Alena Konovalova"]
license = "GNU v3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
redis = "~4.5"
fastapi = "~0.95"
pydantic = { version = "^1.10.8", extras = ["dotenv"] }
orjson = "^3.8.13"
uvloop = "^0.17.0"
elasticsearch = { version = "7.9.1", extras = ["async"] }
uvicorn = "~0.22"
fastapi-cache2 = { version = "~0.2", extras = ["redis"] }
gunicorn = "~20.1"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
opentelemetry-api = "^1.20.0"
opentelemetry-sdk = "^1.20.0"
opentelemetry-instrumentation-fastapi = "^0.41b0"
opentelemetry-exporter-jaeger = "^1.20.0"


[tool.poetry.group.dev.dependencies]
pytest = "7.3.2"
ipython = "^8.13.2"
pre-commit = "^3.3.2"
flake8 = "^6.0.0"
isort = "^5.12.0"
black = "^23.3.0"
types-redis = "^4.5.5.2"
pyproject-flake8 = "^6.0.0.post1"
dlint = "^0.14.1"
flake8-2020 = "^1.8.0"
flake8-bugbear = "^23.5.9"
flake8-builtins = "^2.1.0"
flake8-comprehensions = "^3.12.0"
flake8-deprecated = "^2.0.1"
flake8-variables-names = "^0.0.5"
mccabe = "^0.7.0"
pep8-naming = "^0.13.3"
flake8-pyproject = "^1.2.3"
mypy = "^1.3.0"
pytest-asyncio = "0.21.0"
aiohttp = "3.8.4"
httpx = "^0.25.0"
pytest-mock = "^3.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
